Extract new Corvus.AzureFunctions library
This begins the process, moving the FunctionConfiguration class and
copying the appropriate pieces from the Corvus.SpecFlow.Extensions
project file to the new library's project file.
