+++ Solutions/Corvus.AzureFunctions/Corvus.AzureFunctions.csproj
@@ -19,16 +19,6 @@
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageReference>
-    <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="2.2.0" />
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="2.2.0" />
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="2.2.0" />
-    <PackageReference Include="NUnit" Version="3.12.0" />
-    <PackageReference Include="SpecFlow" Version="3.0.225" />
-    <PackageReference Include="System.Management" Version="4.7.0" />
-  </ItemGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\Corvus.AzureFunctions\Corvus.AzureFunctions.csproj" />
   </ItemGroup>
-
 </Project>
