+++ Solutions/Corvus.AzureFunctions/Corvus/AzureFunctions/FunctionConfiguration.cs
@@ -2,7 +2,7 @@
 // Copyright (c) Endjin Limited. All rights reserved.
 // </copyright>
 
-namespace Corvus.SpecFlow.Extensions
+namespace Corvus.AzureFunctions
 {
     using System.Collections.Generic;
